+++ framework/db/Expression.php
@@ -28,7 +28,7 @@ namespace yii\db;
  * @author Qiang Xue <qiang.xue@gmail.com>
  * @since 2.0
  */
-class Expression extends \yii\base\BaseObject
+class Expression extends \yii\base\BaseObject implements ExpressionInterface
 {
     /**
      * @var string the DB expression
@@ -57,10 +57,19 @@ class Expression extends \yii\base\BaseObject
 
     /**
      * String magic method.
-     * @return string the DB expression
+     * @return string the DB expression.
      */
     public function __toString()
     {
         return $this->expression;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function buildUsing(QueryBuilder $queryBuilder, &$params = [])
+    {
+        $params = array_merge($params, $this->params);
+        return $this->__toString();
     }
 }