+++ framework/db/mysql/QueryBuilder.php
@@ -10,7 +10,7 @@ namespace yii\db\mysql;
 use yii\base\InvalidParamException;
 use yii\base\NotSupportedException;
 use yii\db\Exception;
-use yii\db\Expression;
+use yii\db\ExpressionInterface;
 
 /**
  * QueryBuilder is the query builder for MySQL databases.
@@ -247,11 +247,8 @@ class QueryBuilder extends \yii\db\QueryBuilder
         } else {
             foreach ($columns as $name => $value) {
                 $names[] = $schema->quoteColumnName($name);
-                if ($value instanceof Expression) {
-                    $placeholders[] = $value->expression;
-                    foreach ($value->params as $n => $v) {
-                        $params[$n] = $v;
-                    }
+                if ($value instanceof ExpressionInterface) {
+                    $placeholders[] = $value->buildUsing($this, $params);
                 } elseif ($value instanceof \yii\db\Query) {
                     list($sql, $params) = $this->build($value, $params);
                     $placeholders[] = "($sql)";