+++ tests/framework/db/pgsql/QueryBuilderTest.php
@@ -7,7 +7,11 @@
 
 namespace yiiunit\framework\db\pgsql;
 
+use yii\db\Expression;
+use yii\db\pgsql\ArrayExpression;
+use yii\db\Query;
 use yii\db\Schema;
+use yiiunit\data\base\TraversableObject;
 
 /**
  * @group db
@@ -76,6 +80,31 @@ class QueryBuilderTest extends \yiiunit\framework\db\QueryBuilderTest
             [['not ilike', 'name', ['heyho', 'abc']], '"name" NOT ILIKE :qp0 AND "name" NOT ILIKE :qp1', [':qp0' => '%heyho%', ':qp1' => '%abc%']],
             [['or ilike', 'name', ['heyho', 'abc']], '"name" ILIKE :qp0 OR "name" ILIKE :qp1', [':qp0' => '%heyho%', ':qp1' => '%abc%']],
             [['or not ilike', 'name', ['heyho', 'abc']], '"name" NOT ILIKE :qp0 OR "name" NOT ILIKE :qp1', [':qp0' => '%heyho%', ':qp1' => '%abc%']],
+
+            // array condition corner cases
+            [['@>', 'id', new ArrayExpression(1)], '"id" @> ARRAY[:qp0]', [':qp0' => 1]],
+            [['@>', 'id', new ArrayExpression(false)], '"id" @> ARRAY[:qp0]', [':qp0' => false]],
+            [['&&', 'price', new ArrayExpression([12, 14], 'float')], '"price" && ARRAY[:qp0, :qp1]::float[]', [':qp0' => 12, ':qp1' => 14]],
+            [['@>', 'id', new ArrayExpression([2, 3])], '"id" @> ARRAY[:qp0, :qp1]', [':qp0' => 2, ':qp1' => 3]],
+            'array of arrays' => [['@>', 'id', new ArrayExpression([[1,2], [3,4]], 'float[][]')], '"id" @> ARRAY[ARRAY[:qp0, :qp1], ARRAY[:qp2, :qp3\\]\\]::float[][]', [':qp0' => 1, ':qp1' => 2, ':qp2' => 3, ':qp3' => 4]],
+            [['@>', 'id', new ArrayExpression([])], '"id" @> \'{}\'', []],
+            'array can not contain nulls' => [['@>', 'id', new ArrayExpression([null])], '"id" @> \'{}\'', []],
+            [['@>', 'id', new ArrayExpression(new TraversableObject([1, 2, 3]))], '[[id]] @> ARRAY[:qp0, :qp1, :qp2]', [':qp0' => 1, ':qp1' => 2, ':qp2' => 3]],
+            [['@>', 'time', new ArrayExpression(new Expression('now()'))], '[[time]] @> ARRAY[now()]', []],
+            [['@>', 'time', new ArrayExpression([new Expression('now()')])], '[[time]] @> ARRAY[now()]', []],
+            [['@>', 'id', new ArrayExpression((new Query())->select('id')->from('users')->where(['active' => 1]))], '[[id]] @> ARRAY(SELECT [[id]] FROM [[users]] WHERE [[active]]=:qp0)', [':qp0' => 1]],
+            [['@>', 'id', new ArrayExpression([(new Query())->select('id')->from('users')->where(['active' => 1])], 'integer')], '[[id]] @> ARRAY[ARRAY(SELECT [[id]] FROM [[users]] WHERE [[active]]=:qp0)::integer[]]::integer[]', [':qp0' => 1]],
+
+            // Checks to verity that operators work correctly
+            [['@>', 'id', new ArrayExpression([1])], '"id" @> ARRAY[:qp0]', [':qp0' => 1]],
+            [['<@', 'id', new ArrayExpression([1])], '"id" <@ ARRAY[:qp0]', [':qp0' => 1]],
+            [['=', 'id',  new ArrayExpression([1])], '"id" = ARRAY[:qp0]', [':qp0' => 1]],
+            [['<>', 'id', new ArrayExpression([1])], '"id" <> ARRAY[:qp0]', [':qp0' => 1]],
+            [['>', 'id',  new ArrayExpression([1])], '"id" > ARRAY[:qp0]', [':qp0' => 1]],
+            [['<', 'id',  new ArrayExpression([1])], '"id" < ARRAY[:qp0]', [':qp0' => 1]],
+            [['>=', 'id', new ArrayExpression([1])], '"id" >= ARRAY[:qp0]', [':qp0' => 1]],
+            [['<=', 'id', new ArrayExpression([1])], '"id" <= ARRAY[:qp0]', [':qp0' => 1]],
+            [['&&', 'id', new ArrayExpression([1])], '"id" && ARRAY[:qp0]', [':qp0' => 1]],
         ]);
     }